requirements.txt
asgiref==3.3.1
author==0.0.0.0.dev20200926
Django==3.1.7
django-ckeditor==6.0.0
django-email==0.1.10
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
